+++ CHANGELOG.rst
@@ -1,3 +1,24 @@
+0.4.0 (2023-04-04)
+==================
+
+Backwards Incompatible Changes
+------------------------------
+
+- This removes the older verison of `sunkit_instruments.goes_xrs.calculate_temperature_em` that no longer works for the re-processed netcdf files and new GOES-R data.
+
+  This also removes the `sunkit_instruments.goes_xrs.calculate_radiative_loss_rate` and `sunkit_instruments.goes_xrs.calculate_xray_luminosity` functions that also no longer work in their current form.
+
+  The new `sunkit_instruments.goes_xrs.calculate_temperature_em` function now returns a new sunpy.timeseries.GenericTimeSeries with the temperature and emission measure rather than appending columns to the passed XRSTimeSeries. (`#81 <https://github.com/sunpy/sunkit-instruments/pull/81>`__)
+
+
+Features
+--------
+
+- Create new function (`sunkit_instruments.goes_xrs.calculate_temperature_em`) to calculate the temperature and emission measure from the GOES XRS measurements including the new re-processed XRS data and GOES-R data.
+
+  See :ref:`sphx_glr_generated_gallery_calculate_goes_temperature_and_emission_measure.py` for an example. (`#81 <https://github.com/sunpy/sunkit-instruments/pull/81>`__)
+
+
 0.3.0 (2022-05-24)
 ==================